- working on relative backgrounds
git-svn-id: svn://localhost/private/Console2/trunk@319 8ae08d04-e83e-804b-908e-3db21359c104

+++ Console2/Console/Cursors.cpp
@@ -1026,7 +1026,7 @@ void FadeBlockCursor::PrepareNext() {
 
 	if (m_blendFunction.SourceConstantAlpha < ALPHA_STEP) {
 		m_nStep = ALPHA_STEP;
-	} else if (static_cast<DWORD>(m_blendFunction.SourceConstantAlpha) + ALPHA_STEP > 255) {
+	} else if (m_blendFunction.SourceConstantAlpha + ALPHA_STEP > 255) {
 		m_nStep = -ALPHA_STEP;
 	}
